dcbf/requirement.txt
# Base Python dependencies required for the OCBF CLI and the tested sampling workflow.
# Tested with Python 3.10.
# Install with:
#   python -m pip install -r requirement.txt

maml==2024.6.13
matplotlib==3.10.8
path==16.14.0
ase==3.28.0
spglib==2.7.0
dscribe==2.1.2
Cython==3.0.11
numpy==2.2.6
pyyaml==6.0.2
pip==24.2
pandas==2.3.3
scipy==1.15.3
scikit-learn==1.7.2
tqdm==4.67.3
loguru==0.7.3
pybind11==3.0.3
git+https://github.com/bigd4/PyNEP.git

# Optional extra backend:
# ABACUS backend:
#   python -m pip install git+https://gitlab.com/1041176461/ase-abacus.git
